2024 update to python3.10 with many libs.

+++ sciwx/__init__.py
@@ -7,7 +7,7 @@ for i in plt.colormaps()[::-1]:
 	cm = plt.get_cmap(i)
 	if i[-2:]=='_r': continue
 	vs = np.linspace(0, cm.N, 256, endpoint=False)
-	lut = cm(vs.astype(np.int), bytes=True)[:,:3]
+	lut = cm(vs.astype(np.uint8), bytes=True)[:,:3]
 	ColorManager.add(i, lut)
 del plt
 graylut = ColorManager.get('gray')
